backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
 
# LangChain — pinned to 0.3.x series (1.x removed langchain.chains)
langchain==0.3.25
langchain-core==0.3.84
langchain-community==0.3.24
langchain-groq==0.2.4
langchain-qdrant==0.2.0
langchain-cohere==0.3.4
 
# Vector DB + Embeddings
qdrant-client
cohere
 
# Monitoring + Utilities
langsmith
gitpython
python-dot